slight cleanup
