--- jetty-impl/src/main/java/works/hop/jetty/HelloServlet.java
@@ -1,17 +1,0 @@
-package works.hop.jetty;
-
-import javax.servlet.http.HttpServlet;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-import java.io.IOException;
-
-public class HelloServlet extends HttpServlet {
-
-    @Override
-    protected void doGet(HttpServletRequest request, HttpServletResponse response) throws IOException {
-        response.setStatus(HttpServletResponse.SC_OK);
-        response.setContentType("text/html");
-        response.setCharacterEncoding("utf-8");
-        response.getWriter().println("<h1>Hello from HelloServlet</h1>");
-    }
-}
