+++ jetty-impl/src/main/java/works/hop/jetty/hello/HelloHandler.java
@@ -1,4 +1,4 @@
-package works.hop.jetty;
+package works.hop.jetty.hello;
 
 import org.eclipse.jetty.server.Request;
 import org.eclipse.jetty.server.handler.AbstractHandler;
@@ -10,6 +10,7 @@ import java.io.IOException;
 import java.io.PrintWriter;
 
 public class HelloHandler extends AbstractHandler {
+
     final String greeting;
     final String body;
 
@@ -27,21 +28,14 @@ public class HelloHandler extends AbstractHandler {
     }
 
     @Override
-    public void handle(String target,
-                       Request baseRequest,
-                       HttpServletRequest request,
-                       HttpServletResponse response) throws IOException,
-            ServletException {
+    public void handle(String target, Request baseRequest, HttpServletRequest request, HttpServletResponse response) throws IOException, ServletException {
         response.setContentType("text/html; charset=utf-8");
         response.setStatus(HttpServletResponse.SC_OK);
-
         PrintWriter out = response.getWriter();
-
         out.println("<h1>" + greeting + "</h1>");
         if (body != null) {
             out.println(body);
         }
-
         baseRequest.setHandled(true);
     }
 }
